Fix issue about collision
Add replacement at the outside of the paddle when the ball is detect in the
paddle.
Add variable to check if the ball was already in the paddle at the last
iteration to avoid that the ball is lock in the paddle.
Add condition to have logical rebound in the corners.