Rebound is compute related at the paddle and ball
The variable about the speed for axis x is compute from the position of
the rebound of the ball on the paddle.
The variable about the speed for the axis y is compute from a standard
speed and from the speed for axis x to have the same speed in all directions.